+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## [Unreleased](https://github.com/rr-/Tomb1Main/compare/2.4...master)
 - added CHANGELOG.md
 - added fade animations
+- added a vsync option
 - fixed certain inputs skipping too many things
 - fixed a memory leak in the audio sampler
-- added a vsync option
 
 
 ## [2.4](https://github.com/rr-/Tomb1Main/compare/2.3...2.4) - 2022-01-19